requirements.txt
asgiref==3.8.1
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-localflavor==4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
git-filter-repo==2.47.0
gnu==0.0.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-stdnum==1.20
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
